refactor: extract OverviewLive and delete DashboardLive god-module (#904)
Final tab extraction: the `:overview` live_action becomes
KlassHeroWeb.Provider.OverviewLive at /provider/dashboard, owning the KPI stats,
top programs, pending requests/enrollments, business profile card,
approve_enrollment, and — as the only tab that subscribes to PubSub — all three
handle_info clauses plus the session-stats and enrollment-confirmed
subscriptions. The old handle_params(:overview) load folds into mount/3.

With all four tabs extracted, the 2347-LOC provider/dashboard_live.ex is deleted.
The four sub-LiveViews now share only the three helper modules (Params, Uploads,
Chrome) and the pv_dashboard_shell component.
